feat: tanstack zustand 처리 업로드 형식 변경
