+++ package.json
@@ -35,6 +35,7 @@
     "react-kakao-maps-sdk": "^1.1.27",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7",
+    "uuid": "^11.0.2",
     "zod": "^3.23.8",
     "zustand": "^5.0.0"
   },
@@ -52,6 +53,7 @@
     "@storybook/test": "^8.3.6",
     "@storybook/testing-library": "^0.2.2",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@types/file-saver": "^2.0.7",
     "@types/node": "^20.16.13",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.1",
