+++ src/components/post/_post/ImageUploadForm.tsx
@@ -1,9 +1,10 @@
 'use client';
 
-import { useUploadImage } from '@/hooks/queries/byUse/useUploadImage';
-import { downloadAllAsZip, downloadSingleFile } from '@/services/client-action/downloadAction';
-import { useImageUploadStore } from '@/stores/useImageUploadStore';
+import { downloadSingleFile, downloadAllAsZip } from '@/services/client-action/downloadAction';
+import { generateUniqueFileName } from '@/services/client-action/fileActions';
+import { useImageUploadStore } from '@/stores/imageUploadStore';
 import { formatDateToNumber } from '@/utils/dateUtils';
+import { removeFileExtension } from '@/utils/fileNameUtils';
 import browserClient from '@/utils/supabase/client';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
@@ -23,25 +24,96 @@ interface ImageListProps {
   userId: string;
 }
 
-const Test = ({ userId }: ImageListProps) => {
+const ImageUploadForm = ({ userId }: ImageListProps) => {
   const bucketName = 'tour_images';
   const folderName = 'group_name';
+  const currentDate = new Date().toISOString();
   const formattedDate = formatDateToNumber(new Date().toString()) || '';
   const queryClient = useQueryClient();
   const { images, addImages, updateImage, deleteImage } = useImageUploadStore();
 
-  // useUploadImage 훅을 사용하여 이미지 업로드 처리
-  const uploadMutation = useUploadImage(bucketName, folderName, userId);
+  const fetchSignedUrl = async (filename: string) => {
+    const { data, error } = await browserClient.storage
+      .from(bucketName)
+      .createSignedUrl(`${folderName}/${filename}`, 60 * 60);
+    if (error) throw new Error('Signed URL 생성 실패');
+    return data.signedUrl;
+  };
+
+  const uploadMutation = useMutation({
+    mutationFn: async (files: File[]) => {
+      const formData = new FormData();
+      files.forEach((file) => formData.append('photos', file));
+
+      const exifResponse = await fetch('/api/upload', {
+        method: 'POST',
+        body: formData,
+      });
+
+      if (!exifResponse.ok) throw new Error('EXIF 데이터를 가져오는 데 실패했습니다.');
+      const exifDataArray = await exifResponse.json();
+
+      const uploadedImages = await Promise.all(
+        files.map(async (file, index) => {
+          const uniqueFileName = await generateUniqueFileName(file.name, folderName, bucketName);
+          await browserClient.storage.from(bucketName).upload(`${folderName}/${uniqueFileName}`, file);
+          const signedUrl = await fetchSignedUrl(uniqueFileName);
+          const response = await fetch(signedUrl);
+          const blobUrl = URL.createObjectURL(await response.blob());
+          const exifData = exifDataArray[index];
+          const fileNameWithoutExtension = removeFileExtension(uniqueFileName);
+
+          const { data, error } = await browserClient
+            .from('images')
+            .insert({
+              post_image_name: fileNameWithoutExtension,
+              post_image_url: signedUrl,
+              created_at: currentDate,
+              is_cover: false,
+              post_lat: exifData.latitude,
+              post_lng: exifData.longitude,
+              origin_created_at: formatDateToNumber(exifData.dateTaken),
+              user_id: userId,
+            })
+            .select();
+
+          if (error) throw new Error(error.message);
+          return {
+            blobUrl,
+            id: data[0].id,
+            userId: data[0].user_id,
+            isCover: data[0].is_cover,
+            createdAt: currentDate,
+            filename: uniqueFileName,
+            latitude: exifData.latitude,
+            longitude: exifData.longitude,
+            dateTaken: exifData.dateTaken,
+          };
+        }),
+      );
+
+      addImages(uploadedImages);
+      return uploadedImages;
+    },
+    onSuccess: () => queryClient.invalidateQueries({ queryKey: ['images', userId] }),
+    onError: (error) => {
+      console.error('이미지 업로드 에러:', error);
+    },
+  });
+
   const setCoverMutation = useMutation({
     mutationFn: async (id: number) => {
+      // 모든 이미지의 대표 이미지를 해제
       await browserClient.from('images').update({ is_cover: false }).eq('user_id', userId);
+      // 선택한 이미지를 대표 이미지로 설정
       await browserClient.from('images').update({ is_cover: true }).eq('id', id);
     },
     onMutate: async (id) => {
-      useImageUploadStore.getState().updateImage(id, { isCover: true });
+      // UI에서 직접 반영
+      useImageUploadStore.getState().updateImage(id, { isCover: true }); // 현재 선택한 ID가 대표 이미지로 설정
       useImageUploadStore.getState().images.forEach((image) => {
         if (image.id !== id) {
-          useImageUploadStore.getState().updateImage(image.id, { isCover: false });
+          useImageUploadStore.getState().updateImage(image.id, { isCover: false }); // 다른 이미지들은 대표 이미지 해제
         }
       });
     },
@@ -55,10 +127,13 @@ const Test = ({ userId }: ImageListProps) => {
 
   const deleteMutation = useMutation({
     mutationFn: async (id: number) => {
+      // 데이터베이스에서 파일 이름을 가져옵니다.
       const { data, error } = await browserClient.from('images').select('post_image_name').eq('id', id);
       if (error || !data || data.length === 0) throw new Error('이미지 이름을 가져오는 중 오류 발생');
 
       const filename = data[0].post_image_name;
+
+      // 스토리지에서 파일 삭제
       const { error: storageError } = await browserClient.storage
         .from(bucketName)
         .remove([`${folderName}/${filename}`]);
@@ -67,12 +142,14 @@ const Test = ({ userId }: ImageListProps) => {
         throw new Error('스토리지에서 파일 삭제 중 오류 발생');
       }
 
+      // 데이터베이스에서 이미지 삭제
       const { error: dbError } = await browserClient.from('images').delete().eq('id', id);
       if (dbError) throw new Error('데이터베이스에서 이미지 삭제 중 오류 발생');
 
-      return id;
+      return id; // 삭제한 id를 반환하여 onSuccess에서 사용
     },
     onMutate: async (id) => {
+      // 낙관적 업데이트: 상태에서 삭제
       useImageUploadStore.getState().deleteImage(id);
     },
     onError: (error) => {
@@ -93,6 +170,7 @@ const Test = ({ userId }: ImageListProps) => {
       />
       {images.length > 0 && (
         <div className='flex flex-row gap-2 p-3'>
+          {console.log(images)}
           {images.map((image) => (
             <div
               key={image.id}
@@ -126,4 +204,4 @@ const Test = ({ userId }: ImageListProps) => {
   );
 };
 
-export default Test;
+export default ImageUploadForm;
