--- src/hooks/queries/byUse/useSessionQuery.ts
@@ -1,11 +1,0 @@
-'use client';
-
-import { fetchUserSession } from '@/services/client-action/sessionAction';
-import { useQuery } from '@tanstack/react-query';
-
-export const useSessionQuery = () => {
-  return useQuery({
-    queryKey: ['userSession'],
-    queryFn: fetchUserSession,
-  });
-};
