+++ src/hooks/queries/byUse/useUserQuery.ts
@@ -1,0 +1,11 @@
+'use client';
+
+import { fetchUser } from '@/services/client-action/userAction';
+import { useQuery } from '@tanstack/react-query';
+
+export const useUserQuery = () => {
+  return useQuery({
+    queryKey: ['user'],
+    queryFn: fetchUser,
+  });
+};
