--- src/services/client-action/sessionAction.ts
@@ -1,8 +1,0 @@
-import browserClient from '@/utils/supabase/client';
-
-export const fetchUserSession = async () => {
-  const {
-    data: { session },
-  } = await browserClient.auth.getSession();
-  return session;
-};
