+++ src/services/client-action/userAction.ts
@@ -1,0 +1,8 @@
+import browserClient from '@/utils/supabase/client';
+
+export const fetchUser = async () => {
+  const {
+    data: { user },
+  } = await browserClient.auth.getUser();
+  return user;
+};
