--- src/stores/existingStore.ts
@@ -1,15 +1,0 @@
-import { create } from 'zustand';
-
-interface ImageStore {
-  loading: boolean;
-  error: string | null;
-  setLoading: (loading: boolean) => void;
-  setError: (error: string | null) => void;
-}
-
-export const useImageStore = create<ImageStore>((set) => ({
-  loading: false,
-  error: null,
-  setLoading: (loading) => set({ loading }),
-  setError: (error) => set({ error }),
-}));
